% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_vlist_tmp.R
\name{annotate_vlist}
\alias{annotate_vlist}
\title{Title annotate_vlist}
\usage{
annotate_vlist(vlist.new, bangumi_map, imgur = TRUE)
}
\arguments{
\item{vlist.new}{}
}
\value{
vlist.new.anno
}
\description{
annotate necessary info to generate posts,
}
