% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_aid2cid.R
\name{api_aid2cid}
\alias{api_aid2cid}
\title{Title api_aid2cid}
\usage{
api_aid2cid(aid)
}
\arguments{
\item{aid}{}
}
\value{
cid
}
\description{
used to get cid,which is required to generate the url for html5-player
}
\examples{
# https://www.bilibili.com/video/av2441580/
api_aid2cid(2441580)
}
