% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_getbilitags.R
\name{api_getbilitags}
\alias{api_getbilitags}
\title{getbilitags}
\usage{
api_getbilitags(aid = 3653764)
}
\arguments{
\item{aid}{}
}
\value{
list
}
\description{
# use api, be careful with the quests.
}
\examples{
# https://www.bilibili.com/video/av2441580/
getbilitags(aid = 2441580)
}
