% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_getuploads.R
\name{api_getuploads}
\alias{api_getuploads}
\title{api_getuploads}
\usage{
api_getuploads(mid, kw = "")
}
\arguments{
\item{mid}{bilibili member id}

\item{kw}{keyword for filtering results}
}
\value{
a dataframe of all uploaded videos of mid
}
\description{
api_getuploads
}
\examples{
getvlist(5382023,kw="aaa")

}
