% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_getuploads_fp.R
\name{api_getuploads_fp}
\alias{api_getuploads_fp}
\title{Title api_getuploads, only return first page.}
\usage{
api_getuploads_fp(mid, kw = "")
}
\arguments{
\item{mid}{bilibili member id}

\item{kw}{keyword for search}
}
\value{

}
\description{
Title api_getuploads, only return first page.
}
