% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_upload_imgur.R
\name{api_upload_github}
\alias{api_upload_github}
\title{Title api_upload_github}
\usage{
api_upload_github(pic, outdir)
}
\arguments{
\item{pic}{pic url from bilibili response}

\item{outdir}{output dir of images}
}
\value{

}
\description{
Title api_upload_github
}
