% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromJSON_fix.R
\name{fromJSON_fix}
\alias{fromJSON_fix}
\title{Title fromJSON_fix fixe}
\usage{
fromJSON_fix(
  txt,
  simplifyVector = TRUE,
  simplifyDataFrame = simplifyVector,
  simplifyMatrix = simplifyVector,
  flatten = FALSE,
  ...
)
}
\arguments{
\item{txt}{json text}

\item{simplifyVector}{v}

\item{simplifyDataFrame}{v}

\item{simplifyMatrix}{v}

\item{flatten}{v}

\item{...}{other v}
}
\value{
df
}
\description{
Title fromJSON_fix fixe
}
