% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funs.R
\name{get_existing_aid}
\alias{get_existing_aid}
\title{Title get_existing_aid}
\usage{
get_existing_aid(...)
}
\arguments{
\item{...}{list of formatted posts}
}
\value{
aids, character vector
}
\description{
Title get_existing_aid
}
