% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbangumi.R
\name{getbangumi}
\alias{getbangumi}
\title{Title getbangumi deprecated}
\usage{
getbangumi(vector)
}
\arguments{
\item{vector}{character vector containing some, usually titles, if no keywords detected, return "其他"}
}
\value{
a vector of bangumi,
}
\description{
Title getbangumi deprecated
}
