% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funs.R
\name{getyearsdf}
\alias{getyearsdf}
\title{Get years from vector}
\usage{
getyearsdf(vector_character, date_format = "\%y\%m\%d")
}
\arguments{
\item{vector_character}{character vector containing possible date}

\item{date_format}{date format}
}
\value{
vector of date
}
\description{
Get years from vector
}
